+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- Adds test to verify that session container uses overridden functions
 
 ### Added
 - Adds with-go-zero example: https://github.com/supertokens/supertokens-golang/issues/157