+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Adds with-go-zero example: https://github.com/supertokens/supertokens-golang/issues/157
+- UserId Mapping functionality and compatibility with CDI 2.15
+- Adds `CreateUserIdMapping`, `GetUserIdMapping`, `DeleteUserIdMapping`, `UpdateOrDeleteUserIdMappingInfo` functions to supertokens package
+
 
 ## [0.8.2] - 2022-07-18